remove duplicate useAppContext hook
